+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 
 - Fixed bug where the Nexus.cfg was not consulted when reading CHARACTERS MATRIX.
+- Keep state labels when normalising a NEXUS CHARACTERS block.
+- Better support for Morphbank-style references in NOTES.
 
 
 ## [v1.0.0] - 2023-03-08